Consolidate consumer support classes
