+++ lib/jetstream_bridge/consumer/consumer.rb
@@ -7,7 +7,6 @@ require_relative '../core/duration'
 require_relative '../core/logging'
 require_relative '../core/config'
 require_relative '../core/model_utils'
-require_relative 'consumer_config'
 require_relative 'message_processor'
 require_relative 'subscription_manager'
 require_relative 'inbox/inbox_processor'
